Move jsonable_encoder.py to Opik root
